refactor(app): 拆出 App 编排层、会话 Hub、动效/全局事件 hook 与子视图组件
